Fixed some subscript out of range errors caused by using incorrect index on random range

+++ uav/postal_region.h
@@ -4,7 +4,7 @@ using namespace std;
 struct address_metadata
 {
 	string post_code;
-	int x_coord = 0, y_coord = 0;
+	double x_coord = 0, y_coord = 0;
 	string out_code;
 	string in_code;
 	int id = -1;
